% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_1961_literacy}
\alias{census_1961_literacy}
\title{Census 1961 literacy data}
\format{
A tibble with 347 rows and 8 columns:
\describe{
\item{year}{Census year (1961)}
\item{state}{Name of the state}
\item{state_name_harmonized}{Harmonized state name for joining across datasets}
\item{district}{Name of the district}
\item{state_district_code}{Combined state-district code}
\item{literacy_total}{Total literacy rate (percent)}
\item{literacy_male}{Male literacy rate (percent)}
\item{literacy_female}{Female literacy rate (percent)}
}
}
\source{
Jolad, Shivakumar and Singh, Madhav (2026). "Indian Census Data
Collection, 1901-2026: Digitised Subnational Population and Administrative
Datasets." Harvard Dataverse. \doi{10.7910/DVN/ON8CP8}.
}
\usage{
census_1961_literacy
}
\description{
District-level literacy rates from the 1961 Census of India.
}
\keyword{datasets}
