% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_subdistricts_2011}
\alias{census_subdistricts_2011}
\title{Subdistrict Directory with 2011 Census Data}
\format{
A tibble with 7,074 rows and 15 columns:
\describe{
\item{year}{Census year for population data (2011)}
\item{geography}{Geographic level ("subdistrict")}
\item{state}{Name of the state}
\item{state_name_harmonized}{Harmonized state name for joining across datasets}
\item{district}{Name of the district}
\item{subdistrict}{Name of the subdistrict (tehsil/taluka)}
\item{population}{Total population}
\item{males}{Male population}
\item{females}{Female population}
\item{households}{Number of households}
\item{inhabited_villages}{Number of inhabited villages}
\item{uninhabited_villages}{Number of uninhabited villages}
\item{towns}{Number of towns}
\item{area_km2}{Area in square kilometers}
\item{density}{Population density per square kilometer}
}
}
\source{
Jolad and Singh (2026), "Indian Census Data Collection, 1901-2026:
Digitised Subnational Population and Administrative Datasets",
Harvard Dataverse, \doi{10.7910/DVN/ON8CP8}.
}
\usage{
census_subdistricts_2011
}
\description{
Administrative subdistrict (tehsil/taluka) level data linked to 2011 Census
population figures. Based on 2026 administrative boundaries.
}
\keyword{datasets}
