% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{detect_year}
\alias{detect_year}
\title{Detect year from data}
\usage{
detect_year(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Integer year or NULL
}
\description{
Internal function to detect the census year from a data frame.
}
\keyword{internal}
