% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_census_boundaries}
\alias{get_census_boundaries}
\title{Get census boundaries}
\usage{
get_census_boundaries(year, geography = c("state", "district"))
}
\arguments{
\item{year}{Census year. Available: 1941, 1951, 1961, 1971, 1981, 1991, 2001, 2011.}

\item{geography}{Geographic level: "state" or "district".}
}
\value{
An sf object containing the geographic boundaries with harmonized state names.
}
\description{
Retrieve geographic boundaries for Indian states or districts.
}
\examples{
\dontrun{
boundaries <- get_census_boundaries(1971, "district")
ggplot2::ggplot(boundaries) +
  ggplot2::geom_sf()
}

}
