% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_palette}
\alias{get_palette}
\title{Get color palette}
\usage{
get_palette(name, reverse = FALSE)
}
\arguments{
\item{name}{Name of the palette}

\item{reverse}{Logical. Reverse direction?}
}
\value{
A character vector of colors
}
\description{
Get color palettes for maps. Available palettes: "viridis", "red_blue",
"blue_red", "greens", "oranges", "reds", "blues", "purple_green".
}
