% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiacensus-package.R
\docType{package}
\name{indiacensus-package}
\alias{indiacensus}
\alias{indiacensus-package}
\title{indiacensus: Package to access census of India across the years}
\description{
Provides easy access to digitised Census of India data from 1901 to 2011. Includes population time series, primary census abstracts at state and district levels, and administrative directories. Supports retrieval of census variables with optional geographic boundaries for mapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saketlab.github.io/indiacensus/}
  \item \url{https://github.com/saketlab/indiacensus}
  \item Report bugs at \url{https://github.com/saketlab/indiacensus/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketkc@gmail.com}

}
\keyword{internal}
