% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{list_census_geographies}
\alias{list_census_geographies}
\title{List Available Census Geographies}
\usage{
list_census_geographies()
}
\value{
A tibble showing available year-geography combinations.
}
\description{
Show which geographic levels are available for each census year.
}
\examples{
list_census_geographies()
}
