% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{search_census_variables}
\alias{search_census_variables}
\title{Search census variables}
\usage{
search_census_variables(pattern)
}
\arguments{
\item{pattern}{Character string or regex pattern to search.}
}
\value{
A tibble of matching variables.
}
\description{
Search for census variables by keyword in names and labels.
}
\examples{
search_census_variables("population")
search_census_variables("worker")

}
